requirements.txt
keyboard==0.11.0
pynput==1.3.10
win32compat==221.26
win32core==221.36
Pillow==5.0.0